[sensors] Split RgbdSensorDiscrete into its own file
Unless classes are bleed private implementation details to each other,
generally we should place once class per h/cc file pair to de-clutter
development improve build speeds.
